@@ -13,90 +13,77 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.List;
 
 public class RouteAdapter extends RecyclerView.Adapter<RouteAdapter.RouteViewHolder> {
-    private List<Route> routeList;
-    private int lastSelectedPosition = -1;
-    private Context context;
-
-    public class RouteViewHolder extends RecyclerView.ViewHolder {
-        public TextView nameLabel, tv_time;
-        public RadioButton radioRouteSelected;
-        public RelativeLayout route_layout;
-
-        public RouteViewHolder(View view) {
-            super(view);
-            nameLabel =  view.findViewById(R.id.lblRouteName);
-            tv_time =  view.findViewById(R.id.tv_time);
-            radioRouteSelected = view.findViewById(R.id.radioButtonShehiaSelected);
-            route_layout = view.findViewById(R.id.route_layout);
-        }
-    }
+    private final List<Route> routeList;
+    private final Context context;
+    private int selectedPosition = -1;
 
     public RouteAdapter(List<Route> routeList, Context context) {
         this.routeList = routeList;
         this.context = context;
     }
 
-    @Override
-    public RouteAdapter.RouteViewHolder onCreateViewHolder( ViewGroup parent, int viewType){
-        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.route_row_layout, parent, false);
+    public static class RouteViewHolder extends RecyclerView.ViewHolder {
+        TextView nameLabel, tv_time;
+        RadioButton radioRouteSelected;
+        RelativeLayout route_layout;
 
-        return new RouteAdapter.RouteViewHolder(itemView);
+        public RouteViewHolder(View itemView) {
+            super(itemView);
+            nameLabel = itemView.findViewById(R.id.lblRouteName);
+            tv_time = itemView.findViewById(R.id.tv_time);
+            radioRouteSelected = itemView.findViewById(R.id.radioButtonShehiaSelected);
+            route_layout = itemView.findViewById(R.id.route_layout);
+        }
     }
 
     @Override
-    public void onBindViewHolder( final RouteAdapter.RouteViewHolder holder, final int position) {
-        Route routeObject = routeList.get(position);
-        holder.nameLabel.setText(routeObject.getRouteName());
-        holder.tv_time.setText(routeObject.getDeparture_time());
-        holder.radioRouteSelected.setChecked(lastSelectedPosition == position);
-        holder.radioRouteSelected.setTag(routeList.get(position));
+    public RouteViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(context)
+                .inflate(R.layout.route_row_layout, parent, false);
+        return new RouteViewHolder(view);
+    }
 
-        holder.radioRouteSelected.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v){
-                for(Route _route: routeList){
-                    _route.setSelected(false);
-                }
-                lastSelectedPosition = position;
-                RadioButton radioButton = (RadioButton) v;
-                Route shehia = (Route) radioButton.getTag();
-                shehia.setSelected(true);
-                routeList.get(position).setSelected(true);
-                notifyDataSetChanged();
-            }
-        });
+    @Override
+    public void onBindViewHolder(final RouteViewHolder holder, final int position) {
+        Route route = routeList.get(position);
+        System.out.println("Route Adapter Name: " + route.getRouteName());
+        holder.nameLabel.setText(route.getRouteName());
+        holder.tv_time.setText(route.getDeparture_time());
+        holder.radioRouteSelected.setChecked(position == selectedPosition);
 
-        holder.nameLabel.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                for(Route _route: routeList){
-                    _route.setSelected(false);
-                }
-                lastSelectedPosition = position;
-                RadioButton radioButton = holder.radioRouteSelected;
-                Route route = (Route) radioButton.getTag();
-                route.setSelected(true);
-                routeList.get(position).setSelected(true);
-                notifyDataSetChanged();
-            }
-        });
+        View.OnClickListener clickListener = v -> {
+            updateSelection(position);
+        };
 
-        holder.route_layout.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                for(Route _route: routeList){
-                    _route.setSelected(false);
-                }
-                lastSelectedPosition = position;
-                RadioButton radioButton = holder.radioRouteSelected;
-                Route route = (Route) radioButton.getTag();
-                route.setSelected(true);
-                routeList.get(position).setSelected(true);
-                notifyDataSetChanged();
-            }
-        });
+        holder.itemView.setOnClickListener(clickListener);
+        holder.route_layout.setOnClickListener(clickListener);
+        holder.nameLabel.setOnClickListener(clickListener);
+        holder.tv_time.setOnClickListener(clickListener);
+        holder.radioRouteSelected.setOnClickListener(clickListener);
     }
 
     @Override
     public int getItemCount() {
         return routeList.size();
+    }
+
+    private void updateSelection(int position) {
+        if (position == RecyclerView.NO_POSITION)
+            return;
+        if (selectedPosition >= 0 && selectedPosition < routeList.size()) {
+            routeList.get(selectedPosition).setSelected(false);
+            notifyItemChanged(selectedPosition);
+        }
+
+        selectedPosition = position;
+        routeList.get(position).setSelected(true);
+        notifyItemChanged(position);
+    }
+
+    public Route getSelectedRoute() {
+        if (selectedPosition >= 0 && selectedPosition < routeList.size()) {
+            return routeList.get(selectedPosition);
+        }
+        return null;
     }
 }
