@@ -6,63 +6,50 @@ import android.os.VibrationEffect;
 import android.os.Vibrator;
 
 public class Config {
-
-    /* Logs EndPoint */
+    /* Logs Endpoint */
     static String logsEndpoint = "https://172.19.255.4/api/";
 
-    /* Online EndPoint */
-//    static String  app_ip = "https://booking.zff.co.tz/api/scanner/";
+    // static String app_ip = "https://booking.zff.co.tz/api/scanner/";
 
-    /* Demo EndPoint */
-//    static String  app_ip = "https://demo.zanzibarfastferries.com/api/scanner/";
-//    static String  app_ip = "https://ferries.rahisi.co.tz/api/scanner/";
+    // static String app_ip = "https://demo.zanzibarfastferries.com/api/scanner/";
 
-    /* Local EndPoints */
-    static String  app_ip = "http://172.16.10.217:2019/scanner/";   // Vicent
-//    static String  app_ip = "http://172.16.10.164:7800/scanner/";   // John
-//    static String logsEndpoint = "http://172.16.10.143:8052/api/";
+    // static String app_ip = "https://ferries.rahisi.co.tz/api/scanner/";
 
+    // static String app_ip = "http://172.16.10.217:2019/scanner/";  // Vicent
+    static String app_ip = "http://172.16.10.171:7800/scanner/";     // John
+
+    /* API Routes */
+    public static final String SPLASH_URL = app_ip + "splash";
     public static final String LOGIN_URL = app_ip + "scanner_login.php";
     public static final String LOGOUT_URL = app_ip + "scanner_logout.php";
+    public static final String CHANGE_PASSWORD = app_ip + "api/change_password";
+    public static final String RESET_PASSWORD = app_ip + "api/reset_password";
     public static final String BOOKING_URL = app_ip + "scanner_fetch_bookings.php";
     public static final String UPDATE_TICKET_URL = app_ip + "customerboardingchecking";
     public static final String FIND_TICKET_URL = app_ip + "scanner_get_ticket.php";
     public static final String VERIFY_TICKET_URL = app_ip + "scanner_verify_ticket.php";
     public static final String REPORT_ERROR_LOG = logsEndpoint + "issues_reporting/error_reporting.php";
 
-
     public static final String API_USER_NAME = "goandroy";
     public static final String API_PASSWORD = "12345";
     public static final String API_KEY = "D6H8SKKRL79RJ4WWP5LASYNGJ";
     public static final String API_KEY_NAME = "API-KEY";
-
-    //Keys for operator id and password as defined in our $_POST['key'] in login.php
     public static final String KEY_OPERATOR = "operator";
-    public static final String KEY_PASSWORD = "password";
-
-    //Keys for operator id and password as defined in our $_POST['key'] in login.php
     public static final String KEY_OPERATOR_NAME = "operator_name";
     public static final String KEY_STATUS = "status";
-    public static final String USER_PORT_NAME = "user_port_id";
-    public static final String USER_DEVICE_NAME = "user_device_id";
-
-    //If server response is equal to this that means login is successful
     public static final String LOGIN_SUCCESS = "success";
     public static final String LOGOUT_SUCCESS = "You have successfully logged out";
-
-    //Keys for Sharedpreferences
-    //This would be the name of our shared preferences
     public static final String SHARED_PREF_NAME = "operator_login";
-    //This would be used to store the operator id of current logged in user
-    public static final String OPERATOR_SHARED_PREF = "operator";
-    public static final String ROUTE_SHARED_PREF = "route";
-    //We will use this to store the boolean in sharedpreference to track user is loggedin or not
+    public static final String ROUTE_SHARED_PREF = "route_prefs";
     public static final String LOGGEDIN_SHARED_PREF = "loggedin";
-    // Used to store the name of the currently logged in operator
+    public static final String OPERATOR_SHARED_PREF = "operator";
     public static final String OPERATOR_NAME_SHARED_PREF = "operator_name";
-    public static final String OPERATOR_TYPE_SHARED_PREF = "operator_type";
-
-    // OFFLINE DATABASE CONFIGURATIONS
+    public static final String SHARED_PREF_NAME_ROUTES = "zff_prefs";
+    public static final String PREF_ROUTES_JSON = "routes_json";
+    public static final String PREF_SELECTED_ROUTE_ID = "selected_route_id";
+    public static final String PREF_SELECTED_ROUTE_NAME = "selected_route_name";
+    public static final String PREF_SELECTED_ROUTE_TIME = "selected_route_time";
+    public static final String PREF_SELECTED_ACTION = "selected_action";
     public static final String TICKET_DATABASE_NAME = "zanfastTicketsDb";
     public static final String TICKET_TABLE_NAME = "zanfastTikets";
     public static final int TICKET_DATABASE_VERSION = 3;
@@ -78,26 +65,20 @@ public class Config {
     public static final String TICKET_STATUS = "status";
     public static final String RESPONSE_CODE = "code";
     public static final String TICKET_SCAN_COUNT = "scan_count";
-
     public static final String SHARED_TOTAL_TICKETS = "total_tickets";
     public static final String SHARED_SCANNED_TICKETS = "scanned_tickets";
     public static final String SHARED_DEPARTURE_DATE = "departure_date";
-    public static final String SHARED_USER_PORT_NAME = "user_port_id";
-    public static final String SHARED_USER_DEVICE_NAME = "user_device_id";
-
     public static final String DEFAULT_LANGUAGE = "en";
 
-    public  static String getSerialNo(){
-        String serial = Build.SERIAL;
-        return serial;
+    public static String getSerialNo() {
+        return Build.SERIAL;
     }
 
-    public static void doVibration(Context context){
+    public static void doVibration(Context context) {
         Vibrator v = (Vibrator) context.getSystemService(Context.VIBRATOR_SERVICE);
-        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O){
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             v.vibrate(VibrationEffect.createOneShot(500, VibrationEffect.DEFAULT_AMPLITUDE));
-//            v.vibrate(VibrationEffect.createWaveform(new long[]{0, 500, 1000, 500}, -1));
-        }else{
+        } else {
             v.vibrate(500);
         }
     }
